Invoice-Search/backend/requirements.txt
fastapi
uvicorn
pydantic
azure.identity
requests
opencensus-ext-azure
pytz
